+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.6.0] - 2024-12-17
+### Fixed
+- [#79] Restart and unhealthy component operator after it is installed
+### Changed
+- [#79] Set version in component before installation when version is empty.
+
 ## [v1.5.0] - 2024-12-12
 ### Added
 - [#77] Add labels with component name and version also to pod-templates of Deployments, StatefulSets, DaemonSets, Jobs & CronJobs